Use floats for add_hours

+++ cogs/timeclock.py
@@ -193,7 +193,10 @@ class timeclock:
 
     @clock.command()
     async def add_hours(
-        self, ctx: commands.Context, hours: int, member: Optional[discord.Member] = None
+        self,
+        ctx: commands.Context,
+        hours: float,
+        member: Optional[discord.Member] = None,
     ):
         """Add hours to members (or yourself, if no member is specified.)"""
         if member is not None:
